requirements.txt
# Zaytrics Smart Crowd Monitoring System - Dependencies
# Python 3.8+ required

# Core ML/CV Libraries
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1

# YOLOv8 and Detection
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0

# Web Framework
flask==3.0.0

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0

# Utilities
pandas==2.1.1
pyyaml==6.0.1
tqdm==4.66.1

# Optional GPU support (comment out if CPU only)
# Install manually if needed:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
